+++ src/components/game/game.js
@@ -80,3 +80,18 @@ export function updateGameSettingModalMultilingualContent() {
     }
   }
 }
+
+export function updateScoreModalContent() {
+  document.getElementById('scoreModalLabel').innerHTML = i18next.t('scoreModalLabel');
+  document.getElementById('win-label').innerHTML = i18next.t('win-label');
+  document.getElementById('win-label').innerHTML = i18next.t('win-label');
+  document.getElementById('lose-label').innerHTML = i18next.t('lose-label');
+  document.getElementById('save-score').innerHTML = i18next.t('save-score');
+
+  document.getElementById('email2faModalLabel').innerHTML = i18next.t('email2faModalLabel');
+  document.getElementById('emailAddressLabel').innerHTML = i18next.t('emailAddressLabel');
+  document.getElementById('emailInput').placeholder = i18next.t('emailInput');
+  document.getElementById('send-email-code-button').innerHTML = i18next.t('send-email-code-button');
+  document.getElementById('codeInputLabel').innerHTML = i18next.t('codeInputLabel');
+  document.getElementById('send-verification-code-button').innerHTML = i18next.t('send-verification-code-button');
+}
