+++ src/pages/game.js
@@ -101,8 +101,6 @@ function createGameSettingModal() {
   return modalContainer;
 }
 
-
-
 export function pageBoard() {
   const page = document.createElement('div');
   page.setAttribute('class', 'board');
@@ -329,20 +327,7 @@ function createEmail2faModal() {
 
   return email2faModal;
 }
-export function updateScoreModalContent() {
-  document.getElementById('scoreModalLabel').innerHTML = i18next.t('scoreModalLabel');
-  document.getElementById('win-label').innerHTML = i18next.t('win-label');
-  document.getElementById('win-label').innerHTML = i18next.t('win-label');
-  document.getElementById('lose-label').innerHTML = i18next.t('lose-label');
-  document.getElementById('save-score').innerHTML = i18next.t('save-score');
 
-  document.getElementById('email2faModalLabel').innerHTML = i18next.t('email2faModalLabel');
-  document.getElementById('emailAddressLabel').innerHTML = i18next.t('emailAddressLabel');
-  document.getElementById('emailInput').placeholder = i18next.t('emailInput');
-  document.getElementById('send-email-code-button').innerHTML = i18next.t('send-email-code-button');
-  document.getElementById('codeInputLabel').innerHTML = i18next.t('codeInputLabel');
-  document.getElementById('send-verification-code-button').innerHTML = i18next.t('send-verification-code-button');
-}
 
 function formatCurrentTime() {
   const now = new Date();
